requirements.txt
Django==2.2.7
music-metadata-territories>=19rc1
git+git://github.com/musicmetadata/edi@master
pytz==2019.3
whitenoise
gunicorn

